+++ src/array/items.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable import/extensions */
 import { createValidator } from '../createValidator/createValidator'
-import { minCharactersYobta } from '../minCharactersYobta'
 import { requiredYobta } from '../requiredYobta'
+import { minCharacters } from '../string/minCharacters'
 import { string } from '../string/string'
 import { array } from './array'
 import { items } from './items'
@@ -9,7 +9,11 @@ import { items } from './items'
 const validate = createValidator(
   requiredYobta(),
   array(),
-  items(requiredYobta(), string(), minCharactersYobta(5)),
+  items(
+    requiredYobta(),
+    string(),
+    minCharacters(() => 5),
+  ),
 )
 
 it('accepts empty array', () => {