+++ src/safe/safe.test.ts
@@ -1,12 +1,17 @@
 /* eslint-disable import/extensions */
 import { createValidator } from '../createValidator/createValidator'
-import { minCharactersYobta } from '../minCharactersYobta'
 import { requiredYobta } from '../requiredYobta'
 import { string } from '../string'
+import { minCharacters } from '../string/minCharacters'
 import { safe } from './safe'
 
 const validate = createValidator(
-  safe('catched yobta!', string(), requiredYobta(), minCharactersYobta(5)),
+  safe(
+    'catched yobta!',
+    string(),
+    requiredYobta(),
+    minCharacters(() => 5),
+  ),
 )
 
 it('does not catch when no errors', () => {