+++ src/string/index.ts
@@ -1,4 +1,5 @@
 export * from './email.js'
 export * from './maxCharacters.js'
+export * from './minCharacters.js'
 export * from './reEmail.js'
 export * from './string.js'