+++ src/string/maxCharacters.test.ts
@@ -4,7 +4,10 @@ import { createValidator } from '../createValidator/createValidator'
 import { maxCharacters, maxCharactersMessage } from './maxCharacters'
 
 const customMessage = (limit: number): string => `${limit} yobta!`
-const validate = createValidator(string(), maxCharacters(1, customMessage))
+const validate = createValidator(
+  string(),
+  maxCharacters(() => 1, customMessage),
+)
 
 it('accepts exact lenght', () => {
   const result = validate('a')
@@ -22,7 +25,10 @@ it('regects greater lenght', () => {
 })
 
 it('has default error message', () => {
-  const validateDefault = createValidator(string(), maxCharacters(1))
+  const validateDefault = createValidator(
+    string(),
+    maxCharacters(() => 1),
+  )
   const attempt = (): any => validateDefault('ab')
   expect(attempt).toThrow(maxCharactersMessage(1))
 })